fold changes into existing controller, remove uneeded connector
